@@ -11,44 +11,44 @@
   "author": "Dennis Pfisterer <github@farberg.de>",
   "license": "BSD",
   "dependencies": {
-    "@awaitjs/express": "^0.8.0",
+    "@awaitjs/express": "^0.9.0",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "body-parser": "^1.19.0",
-    "commander": "^8.2.0",
+    "body-parser": "^1.19.1",
+    "commander": "^8.3.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-session": "^1.17.2",
     "formik": "^2.2.9",
     "formik-material-ui": "^3.0.1",
     "js-yaml": "^4.1.0",
-    "keycloak-connect": "^15.0.2",
+    "keycloak-connect": "^15.1.0",
     "kubernetes-client": "^9.0.0",
     "log4js": "^6.3.0",
     "react": "^17.0",
     "react-dom": "^17.0",
     "react-promise-tracker": "^2.1.0",
-    "react-router-dom": "^5.3.0",
+    "react-router-dom": "^6.1.1",
     "react-spinners": "^0.11.0",
-    "use-http": "^1.0.24"
+    "use-http": "^1.0.26"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/preset-env": "^7.15.6",
-    "@babel/preset-react": "^7.14.5",
-    "@svgr/webpack": "^5.5.0",
-    "babel-loader": "^8.2.2",
-    "browser-sync": "^2.27.5",
+    "@babel/core": "^7.16.5",
+    "@babel/plugin-proposal-class-properties": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
+    "@babel/preset-react": "^7.16.5",
+    "@svgr/webpack": "^6.1.2",
+    "babel-loader": "^8.2.3",
+    "browser-sync": "^2.27.7",
     "connect-browser-sync": "^2.1.0",
-    "css-loader": "^6.3.0",
-    "html-loader": "^2.1.2",
-    "html-webpack-plugin": "^5.3.2",
-    "nodemon-webpack-plugin": "^4.5.2",
+    "css-loader": "^6.5.1",
+    "html-loader": "^3.0.1",
+    "html-webpack-plugin": "^5.5.0",
+    "nodemon-webpack-plugin": "^4.7.0",
     "shebang-loader": "0.0.1",
-    "style-loader": "^3.2.1",
-    "webpack": "^5.53.0",
-    "webpack-cli": "^4.8.0",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
     "webpack-node-externals": "^3.0.0"
   },
   "babel": {
